requirements.txt
# Document parsing
python-docx>=1.1.0
docx2txt>=0.8
textract>=1.6.5
pypdf2>=3.0.0

# NLP
spacy>=3.7.0
https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.8.0/pt_core_news_lg-3.8.0-py3-none-any.whl

# Data generation
faker>=28.0.0

# Web interface
streamlit>=1.38.0

# CLI
typer>=0.12.0
rich>=13.0.0

# Cryptography
cryptography>=43.0.0

# Utilities
python-dateutil>=2.9.0
watchdog